Create DrivingCosts
Driving is expensive. Write a program with a car's miles/gallon and gas dollars/gallon (both floats) as input, and output the gas cost for 10 miles, 50 miles, and 400 miles. Output the gas cost (gasCost) with two digits after the decimal point, which can be achieved as follows:
Put gasCost to output with 2 decimal places