Create MusicalNoteFrequencies
On a piano, a key has a frequency, say f0. Each higher key (black or white) has a frequency of f0 * rn, where n is the distance (number of keys) from that key, and r is 2(1.0/12.0). Given an initial key frequency, output that frequency and the next 4 higher key frequencies. Output all frequencies with five digits after the decimal point, which can be achieved as follows:
Put frequency to output with 5 decimal places