Create SeasonDetermination
Write a program that takes a date as input and outputs the date's season. The input is an integer to represent the month and an integer to represent the day. In addition, check if both integers are valid (an actual month and day).

The dates for each season are:
spring: March 20 - June 20
summer: June 21 - September 21
autumn: September 22 - December 20
winter: December 21 - March 19